Initial implementation of Basic.lseries().
lseries is a lazy series, which returns a generator yielding term by term of
the series (thus never ending if the series is infinite).

Basic._eval_lseries() contains a general implementation using nseries() which
is very slow, but works.

Subclasses should implement it's _eval_lseries(), that can be very optimized,
because the implementation should just use .lseries() to get only terms that it
needs.

This patch implements lazy series for Basic, Function, exp, Integral and
Derivative. In addition it also implements Integral._eval_nseries(), that is
needed in Integral lazy evaluation. Tests were written for some basic cases.

This is by no means a complete implementation, just a start.

Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>