seterr() implemented, just like in NumPy.
This allows setting what happens on 0/0. Either an exception is raised, or nan
returned. The default is to return nan, as it used to be.

Having the option to return an exception is extremely useful for debugging,
when you get nan, but you want to know when exactly something went wrong. So
you just do seterr(True) and rerun your code.

Note: when you do seterr(True), not everything is working properly, for
example:

In [1]: integrate(log(x), (x, 0, 1))
Out[1]: -1

restart isympy (due to caching):

In [1]: seterr(True)

In [2]: integrate(log(x), (x, 0, 1))
Exception: Indeterminate 0/0

So seterr() is default set to False. There is a question, whether our
algorithms should rely on the 0/0 behavior.

Signed-off-by: Riccardo Gori <goriccardo@gmail.com>
Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>