Simplify result of spherical harmonics with new exponentiation rules
We assume that the argument th is in (0,pi) and therefore sin(th) > 0
this allows some simplifications to the result.

This patch is intended to keep current behaviour unchanged after applying
the exponentiation patch, so the current tests should be enough.
It doesnt change the results when applied to current sympy.