[1/2] implement correct rules for simplifying exponentiation
Currently sympy gives inconsistent results when simplifying exponentiation.
For example

>>> sqrt(1/x)
x**(-1/2)
>>> 1/sqrt(x)
x**(-1/2)

but these are in fact not equal:

>>> sqrt(1/(-1))
I
>>> 1/sqrt(-1)
-I

also these results are in conflict:

>>> (x**3)**Rational(1,3)
x
>>> ((-1)**3)**Rational(1,3)
(-1)**(1/3)

as principal value of (-1)**(1/3) != -1.
This is because (x**a)**b is not equal to x**(a*b) in general.
For x complex and a,b real, the equation (x**a)**b == x**(a*b) holds
if and only if
modpi(b * modpi(a * arg(x))) == modpi(a * b * arg(x))
where modpi(x) = ((x + pi) mod pi) - pi

The above issues are solved by this patch.
Only the most general cases of this rule are implemented in this patch, in
order not to degrade preformance. Results of exponentiation of exponentiation
should now be always correct.

Implementing these rules causes some tests to fail as they rely on the wrong
behaviour. Next patch in series fixes these tests.