sympy.test() implemented.
This was achieved by reimplementing the gist of py.test in
sympy/utilities/runtests.py. It uses the py.test like reporting and there is a
utility script "bin/test" with a similar usage as py.test.

It is able to run the whole sympy testsuite including XFail, XPass, skip,
and 'raises'. For some little things, the test suite needs to be adapted, which
is done by the following patches.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Riccardo Gori <goriccardo@gmail.com>