[2/2] Fix tests after fixing the exponentiation rules
Some tests started to pass so the XFAIL is removed and some are fixed.
Test test_nseries is fixed by setting right assumptions and test_ccode
is corrected.
The only problematic test is test_heurisch_symbolic_coeffs which fails
inconsistently. The test is XFAILed and issue 1130 was started to fix this.