Fix broken docstrings (#1147)
The patches implementing the rules for simplifying exponentiation
have broken the doctests. The problem was fixed by setting the right
assumptions. In both examples expression of the form 1/(x**a)
needs to be simplified to x**(-a). In the evalf example, the exponent
is summation index, so it is set to integer. The together example
holds only for positive x and real y.