Implement simplification of reciprocal of power
In the rules for simplifying exponentiation implemented recently, I forgot
to implement one important relation for complex x and y:
1/x**y = x**(-y)
This patch adds this functionality. As a side effect it also fixes the problem
with failing heurisch test in issue 1130 so the xfail is removed.

One test in test_ccode is changed because of this simplification and unneeded
assumptions from docstrings are removed.