Fixes sign(y, x) so that it works in python2.6.
The real problem turned out to be with this code in sign():

    elif not e.has(x):
        f= e.evalf()
        if f > 0:
            return 1
        else:
            return -1

Which obviously utterly fails if "e == Symbol('y')", because then "f == y" and
"f > 0" just compares hashes, that changed in python2.6. The solution is to use
"e.is_positive" instead and everything works just fine.

Thorough tests for this issue committed.