This fixes the (2+pi+x**2).extract_leading_order(x) problem.
The problem was that if there is more than one term to return, we need to be
sure to include both terms in the result. This case was wrongly handled in
Add.extract_leading_order(), which we just fixed.

Then there was a problem that some advanced series tests started to hang after
this patch. As it turned out, sometimes Order(..., x) returns O(2) instead of
O(1). So we fixed that by applying this patch (at 2 different places):

-            if o.expr==1:
+            if o.expr.is_number:

And the final problem was, that by fixing the as_leading_term(), we had to call
collect in Pow._eval_nseries() -- previously the wrong as_leading_term() was
effectively achieving the same thing, but that was just a sheer luck.