Fixes an old bug in sympy/core/evalf.py:do_integral().
prec+15 has to be changed to mp.prec.  It needs to be mp.prec for oscillatory
quadrature because mpmath significantly increases the precision when quadosc is
called.

This has been broken in SymPy all along. It only worked in the
previous tests by accident. The change of quadrature algorithm in
mpmath 0.10 exposed the problem.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Fredrik Johansson <fredrik.johansson@gmail.com>