same syntax for symbols() and var(), new syntax for symbols()
The new syntax for symbols is

>>> x, y, z = symbols('x y z') # or symbols('x,y,z')

instead of

>>> x, y, z = symbols('xyz')

Please note that this is the same behavior as for var().
This has the advantage that symbols('alpha beta') is possible.

Currently the deprecated behavior is default for compatibility. This can be
changed using the keyword each_char=False:

>>> symbols('xyz', each_char=False)
xyz # a single symbol

If there is a space or comma in the string, the new behavior is automatically
triggered (if each_char is not set), so you can use the new syntax immediately.

Furthermore, var() has been fixed to accept keyword arguments.

Tests have been added.

Signed-off-by: Vinzent Steinberg <vinzent.steinberg@gmail.com>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>