fixes the "Ignoring redefinition" warning.
The problem was, that this code:

class f(Function):
    pass

class f(Function):
    pass

generated an "Ignoring redefinition ..." warning in the BasicMeta metaclass,
which is remembering all subclasses of Basic, so that one can access them using
"C.SomeSubclass".  Because this is a valid Python code, the code in BasicMeta
was changed to only remember the latest definition, just like in Python.

Also a more descriptive comment was written to the FunctionClass metaclass.