Do not test examples if they are not installed
Previously, the examples directory was checked for code style. However, when
the user installs sympy using "./setup.py install", the examples are not
installed (at least not in the parent directory of "sympy"), so the result is
that sympy.test() fails.

Now we only check the examples dir if it's present.