tests for Matrix.eigenvects() robustified
The ordering of the result of eigenvects() was platform dependent due to
hashes. That is ok, but when we test it, we need to make the comparison
platform independent (otherwise it fails in jython), so I introduced a
canonicalize() method in the matrix tests. See it's docstring for more info.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>