DeferredVector implementation, allowing lambdification using arbitrary vectors
The idea is to create x = DeferredVector and then use x[0] or x[2]
when building functions for lambdify. Then, the first argument to
lambdify can be a single variable, which will then get indexed.

Further ideas:

1) Should this be a DeferredMatrix instead? This would allow entire
matrices to be used this way.  Attention would need to be paid to
allow maximum compatibility with numpy slicing.

2) Should a DeferredVector (optionally?) specify its length at
instantiation? This could allow for automatic shape checking in the
lambda function.

Note, this has not been tested for use with anything other than
lambdification with numpy yet.