Calculate rational roots in roots() before calling degree-specific formulas.
Since no cubics of quartics formula is used in the determination of rational roots and the algorithm is fast, rational roots are first computed for all polynomials. In some cases this also reduces the degree of the polynomial, thus solving polynomials that where before unsolvable.

As a side-effect, rational roots are returned even with quartics or cubics set to False.