Don't import os globally in the top level module.
This could represent a security risk if sympy was allowed to run in a
'safe' environment in which sympy was the only module allowed to be imported
since you could have access to the os module at sympy.os. Currently
django-sympy works this way, althought it is planned to change.
It is no problem for apps that run on the googleapp engine, since the python
interpreter they run on has no os module.