Removed input upper bound from zzx_heu_gcd
The upper bound, suggested by literature, included size of input
polynomials' coefficients and their degrees. However it seems that
it was set far too low. In fact experiments show that no limit is
needed in SymPy's implementation, because heuristic gcd never hanged
and seems that almost never failed. This only verifies conclusions from
Maple implementation which is used almost for any input there. Of course,
in both cases there is a fallback algorithm (in our case modular gcd).
A future question is how this algorithm will scale in multivariate case.