Fixes the A*f(x).diff(x)*A problem (#1327)
Before:

In [1]: A = Symbol("A", commutative=False)

In [2]: A*f(x).diff(x)*A
Out[2]:
  d
A⋅──(f(x))⋅A
  dx

After:

In [1]: A = Symbol("A", commutative=False)

In [2]: A*f(x).diff(x)*A
Out[2]:
d         2
──(f(x))⋅A
dx

The problem was that Derivative was not set as commutative even if f(x) is
commutative. This was fixed by checking if the argument of the derivative is
commutative and if so, making the whole Derivative commutative too. Then
everything works.

Some tests have to be fixed by making Wild() explicitly not depending on "x".
One doctest had to be robustified.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Freddie Witherden <freddie@witherden.org>