recursive trigsimp() implemented, thanks to Alan Bromborsky
The original trigsimp() was renamed to trigsimp_nontrecursive() and the new
trigsimp() got a new parameter recursive --- it either calls the
trigsimp_nontrecursive(), or apply recursively calls trigsimp_nontrecursive()
on the result from the cse() algorithm.

Tests written for things that couldn't be done before and can be done now.

curvilinear_coordinates.py demo was simplified, as it can now handle the
spherical coordinates just fine.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>