./setup.py test_doc now runs our testsuite
so ./setup.py test now also runs our doctest suite at the end (but only if all
regular tests passed). The blacklist from bin/doctest was moved to
sympy/utilities/runtests.py, so that it's enough to just do

sympy.doctest()

and it will just work.

Of course we should work on fixing sympy and mpmath so that no blacklist is
necessary, but currently the blacklist is needed.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>