solve( x**Rational(1,2) - 2, x) fixed
it used to return [sqrt(2)], but the answer is [4]. There was a bug in the
solver, that noone has noticed so far: it did everything correctly, only at the
very end it should power the result to "m", and it used to power it to "1/m",
which is wrong.

More tests were written for this.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Fabian Seoane <fabian@fseoane.net>