Printing precision fixed all over sympy (#1052)
This was done by implementing a ._settings dictionary in the StrPrinter, the
same way as the LatexPrinter can be customized. So far only the
"full_prec" option is implemented, which can have 3 values:

False ... Strip all trailing zeros
True  ... Always show the full precision
"auto" ... Strip zeros if part of a bigger expression (default)

The evalf() tests have it as True.

Before:

In [1]: 1.5*x
Out[1]: 1.50000000000000⋅x

Now:

In [1]: 1.5*x
Out[1]: 1.5⋅x

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Fredrik Johansson <fredrik.johansson@gmail.com>