refactor sympy.solvers.solvers.guess_solve_strategy
Some code was redundant and was erased. Also corrected a bug
( it wrongly treated 3**x-10 ) and created a test for that.

In the tests, I changed some ocurrences of tsolve by solve. Solve
anyway calls tsolve, and this way we ensure that solve correctly
recognizes the equation as trascendental.

Thanks to smichr for the report.