Added homogeneous_order function
Read the docstrings for more on this.  It returns the order of a
homogeneous function and None if it it is not homogeneous.  A function
f(x,y) is homogeneous or order n if f(xt,yt) == t**n*f(x,y).  This only
works on functions of the type g(f(x), x), which are the types of
functions that dsolve will be passing to it.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>