Serious fix in exact engine
The test was not even correct (I was taking a.diff instead of r[a].diff,
so it would always pass.  I was also taking the wrong diffs anyway.  The
tests worked because they were exact, but any non-exact equation that
fit the form f(x,y)+g(x,y)*y'=0 would have returned a wrong result.
This has now been fixed.  When I get homogeneous in, I will have tests
of several functions of that form that are not exact.

Also added note on future return of RootOf from solve.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>