A new general-purpose logic module.
This is a replacement for sympy.core.logic and will be used
extensively by the query module. The old logic module can be
removed once the old assumption system is removed.

Contents
--------

  sympy.logic.boolalg
    - Operators for boolean algebra on symbols {And: &, Or: |, Not: ~}
    - some useful methods, mainly used to convert expressions to
      conjunctive normal form (CNF)

  sympy.logic.inference
    - Rules for inference in propositional logic. Main method here is
      satisfiable(expr), which checks satisfiability of a propositional
      sentence (for now using dpll algorithm)

  sympy.logic.kb
    - Definition of a knowledge base

  sympy.algorithms
    - some algorithms used by higher order methods (just DPLL for now),
      but in a future we should also have Quine-McCluskey for simplification
      of boolean expressions

  sympy.utilities
    - some utility methods, notably a function to convert cnf file into
      sympy expressions

Comparison with sympy.core.logic
--------------------------------
This new module has a wider range of applicability. It is also better
integrated into the sympy object model

Performance
-----------
No performance studies where done.

Reference: Most of the methods in this module are implementations of algorithms
described in the book "AI: A modern approach", http://aima.cs.berkeley.edu/