Added a documentation document on Gotchas and Pitfalls and simplify module
This is a compilation of things that are often done wrongly by people
new to SymPy, Python, or both.  This also adds basic documentation for
the simplify module (it is just docstrings).

Many examples in gotchas.txt are skipped because doctest didn't test
pretty printing and I wanted them to show up pretty printed.  If anyone
knows how to fix it so they don't have to be skipped, that would be good.
Note that some tests are skipped because the purposely demonstrate
an exception.  For example, one example shows infinite recursion, and I didn't
want to print the massive traceback.

This also has a few minor doc related fixes, including changes to some
simplify docstrings so that they autodoc correctly and removing "collect()"
the simplify docstring, as that function doesn't exist!

Signed-off-by: Ondrej Certik <ondrej@certik.cz>