Call init_printing prior to calling the doctests
Also don't test docs under doc/src/ if python version is less
than 2.5. This is because python2.4 does not implement the
SKIP flag. Other programs (like Django) solve this problem
implementing their own doctest module, but we agreed that it
is not worth, since we are about to drop support for 2.4