Makes bernoulli(1000) work in Sage
The problem was, that mpmath.bernfrac(1000) returned Sage Integers that failed
in our caching, which was caused by mpmath.libintmath.list_primes(1000)
returning Sage Integers, which was caused by the code in mpmath.libintmath that
used Sage's list_primes() instead of the mpmath one if Sage was available.

The easiest fix is to just type the result of mpmath.bernfrac() to int and this
is what this patch does. Maybe there is a better solution to fix mpmath to only
return Python ints, but the current solution gets the job done.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Fabian Pedregosa <fabian@fseoane.net>