neg rational power with negative (rat and int) base fixed
If, in the following expression, a is positive (so that -a is negative)

(-a)**(-p/q) ................. [1]

you cannot simply invert the base and negate the exponent, you have to
watch out for the complex roots of the -1 which are

(-1)**x = cos(180*x)+I*sin(180*x).................[2]

When x is an integer (i.e. q=1), there is no complex part and there
is no problem with inverting the base and changing the sign of the
exponent. When q is not 1, e.g. if x=1/3 then (-1)**(1/3) is
.5 + I*sqrt(3)/2 and [1] would be

(-a)**(1/3) = (-1)**(1/3) * a**(1/3) = (.5 + I*sqrt(3)/2) * a**(1/3)

and all is well. BUT if the 1/3 is changed to -1/3 you have

(-1)**(-1/3) * a**(-1/3)

and if you change those by inverting the base and negating the exponent
you get

(-1)**(1/3) * (1/a)**(1/3)

and this is not right since (-1)**(-1/3) is the complex conjugate of
of (-1)**(-1/3) as you can verify with eq [2]. So what should be used
instead? The complex conjugate of (-1)**(1/3) which is -(-1)**(2/3)
as you can verify by multiplying:

(-1)**(1/3) * -(-1)**(2/3) = -(-1)**(1/3+2/3) = -(-1) = 1

In general, for (-1)**(-p/q) we use -(-1)**(((-p)%q)/q), e.g. if the exponent is-1/3 the new exponent is -1%3/3 = 2/3, or for -7/3 we get -7%3/3 = 2/3, too.