Latex Printing: change from inline=True or False to mode with 4 options
Latex Printing now takes a keyword arguement 'mode' which can be one
of 'plain', 'inline', 'equation', or 'equation*'.  'mode' replaces
'inline'.  There is a deprecation warning if 'inline' is used as a
keyword arguement, but it is handled gracefully for now :

'inline'=True --> 'mode'='inline'
'inline'=False --> 'mode'='equation*'

The tests, docstring, doctests, and tutorial all reflect this change.

Signed-off-by: Vinzent Steinberg <Vinzent.Steinberg@gmail.com>