Added discriminant function for Polys.
The discriminant of a univariate polynomial p of degree n is defined as
n**(n*(n-1)/2)/a_n*resultant(p, p'), where p' is the derivative of p and a_n.
is the leading coefficient of p.

See <http://en.wikipedia.org/wiki/Discriminant> for more info.

This is useful because the discriminant of a polynomial vanishes identically
iff the polynomial has repeated roots.  For example, a quadratic a*x**2 + b*x + c
has repeated roots iff the discriminant, b**2 - 4*a*c is equal to 0.

This also adds tests.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>