Fixes all docs to pass doctests
Now the default printer is always reset after each *.txt file run, so it made
some doctests fail. This patch fixes it by the combination of 2 techniques:

1) for just couple failures, it fixes the doctests
2) if the whole *.txt would have to be rewritten, it sets the sys.displayhook
to sympy.pprint, that way all the test pass fine and everything is explicit.

Note that thanks to the previous patch, the technique 2) doesn't cause any
failures in other files.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Aaron Meurer <asmeurer@gmail.com>