Manually renumber solutions in tests to maintain platform independence.
_renumber() from within constantsimp() has been moved out and renamed
ode_renumber().  It uses Basic._comapre_pretty, which is not platform
independent, so the numbering of the constants from a solution on one
machine may be different than the numbering on another.  The way to test
this is to run ode_renumber() on the solution of any ode of order > 1,
which will canonize the order for that machine.

There is a strange thing that happens sometimes where Basic._compare_pretty
cannot renumber the arguments correctly either way.  For example,
compare (C1 + C2*x)*sin(x*sqrt(2)) + (C3 + C4*x)*cos(x*sqrt(2))
and (C1 + C2*x)*cos(x*sqrt(2)) + (C3 + C4*x)*sin(x*sqrt(2))
on a 64-bit machine.  For this reason, the tests test both the
original solution and the renumbered solution.

Hopefully fix ode tests on 64-bits

Signed-off-by: Ondrej Certik <ondrej@certik.cz>