residue() function implemented
This function can calculate residues. Currently only simple ones, like 1/x or
1/sin(x). Later on we should be able to calculate a residue of any function,
for that we need to improve our series expansion code.

Tests and doctests added. Some tests were marked with TODO markers, so that it
can be grepped later and fixed.

References for ideas how a better algorithm can be improved are in the
docstring for residue().