Made 0th Derivative return the function, even when evaluate=False
The reasoning behind this is that there is no good way to pretty print
this (before this patch, it results in a traceback), because you cannot
access the symbols that do not appear, and pretty printing the function
when it is a Derivative class is misleading.

Tests to follow in next commits.