1664 eigenvector multiplicity
The test in the eigenvector routine is that the length of the basis
(the geometric multiplicity) be the *same as* the algebraic multiplicity,
http://www.ms.uky.edu/~lee/amspekulin/eigenvectors.pdf {see pp. 16, 17}.
The condition perhaps should be that len(basis) <= k. All matrices tests
still pass with that change.

See also, http://www.ee.ic.ac.uk/hp/staff/dmb/matrix/eigen.html#GeomMult

Signed-off-by: Ondrej Certik <ondrej@certik.cz>