1689: sympify(1.0+0J) gives Real(1.0)
There was an inconsistent behavior between sympify(1.0)
and sympify(1.0+0J) that has been resolved. Previously,
sympify would convert any real parts to integers if
they were numerically the same. The behavior now is to
return a real representation: if the user has specified
that a real be sympified then they get a real. Note
that returning a Real() of whatever they gave may or
may not return a Real since Real(1) is Integer(1), but
that is the present behavior of Real and is unrelated
to this issue.

A test (as given in the issue) was added.

Signed-off-by: Chris Smith <smichr@gmail.com>
Signed-off-by: Vinzent Steinberg <Vinzent.Steinberg@gmail.com>