Use expr.func to get the head of expressions.
Use expr.func == some_class instead of isinstance(expr, some_class), and
expr.func instead of type(expr) wherever possible.

The Python class of a sympy object is an implementation detail whereas
the head of an expression (i.e. whether it's an addition, a multiplication, etc.)
is a very important piece of data for symbolic manipulation. The latter is
defined by the attribute 'func'. Using it systematically allows to abstract
the notion of head from the implementation, allowing design changes and paving
the way for possible optimisations and refactorings.