Change move symbols before dict in separatevars parameters list
i.e., separatevars(expr, dict=False, symbols=[]) has been changed to
separatevars(expr, symbols=[], dict=True).  This is because symbols is
indented for the user while dict is mainly intended for internal use.
This makes it possible to do separatevars(x*y + y, (x, y)) without the
use of any magic.  Magic that was previously implemented was removed,
and the docstring updated.