Fix issue 1726
classify_ode() was not returning default=None when the expression has
order 0, which was causing an error.  Also, rearranged code in dsolve so
that it tells you that an expression that is not an ODE in the function
given (order 0) is not an ODE instead of raising NotImplementedError

Tests were added.