1785: expand integrand
.expand(mul=True) *could* be applied to an integrand
before integrating. sqrt(x)*(1+x) was failing because
it can't be made into a Poly (with the current Poly module)
and a test was only being done to see if the integrand was
an Add. In this case, the integrand is a Mul that (with a
shallow expand on multiplied terms) is converted into an Add that
is then easily integrated. HOWEVER, there are some expressions
that can't be integrated unless they aren't expanded. Until
those expressions are identified then one of two paths are
possible:

        - expand everything before integrating and
	  go back and try it without expansion if this fails
	- assume the user knows what they are doing and
	  only expand after a failed attempt.

Not knowing which is the better of the two, I've put the
expansion after the 1st attempt has been made.