1789: primetest changes to isprime and mr_safe
Changes inspired by Christian Oudard and Aaron Meurer in Issue 1789.

The test for numbers up to 10**16 has pseuodoprimes that pass it. An
additional test that n not be one of the passing pseudoprimes is
added. The ranges were modified, too, to try use base sets that had
the lowest stot value (the number of factors of 2 that determine how
many loops on s will be done in _test.

The docstring was also expanded to caution about adding new ranges.

A test of a pseudoprime that would have falsely been reported as
being prime was added.

The fallback number of primes that are tested was increased to 46;
the previous set could only prove primality of numbers that are
341550071728321 or less...and this range is already being tested
in mr_safe. The value of 46 for the number to test is arbitrary;
it will assure that the probability of misreporting something as
prime will be 4**-46 (~= 2 x 10^-28). A test with 20 primes is
believed to test up to 10**36 and 46 will test somewhere above
that but less than 10**139 (Arnault's number).

mr_safe has been made into a hidden function and removed from the
ntheory import. It's just a helper function for isprime().