Add visual option to factorint()
This lets you get a visual factorization of an integer, which is much
easier to read than the dictionary normally returned.

For example:

In [1]: factorint(42**2)
Out[1]: {2: 2, 3: 2, 7: 2}

In [2]: factorint(42**2, visual=True)
Out[2]:
 2  2  2
2 ⋅3 ⋅7

This works by using the evaluate=False flag in Mul and Pow.

I also added information to the docstring of factorint(), including two
nice list comprehensions that will convert between one form and the other.

Tests added.