Write ode_sol_simplicity() to replace compare_ode_sol()
This is a key function replacement, which means it will be faster and
Python 3 compatible.

An analysis of the code showed that if the "all" hint is passed to
dsolve() for an ODE with a complex solution (i.e., one that is not
solvable in f(x), such as an ODE that matches the
1st_homogeneous_coefficients hints), most of the time was being spent in
compare_ode_sol() trying to solve the solutions (a solution that is
unsolved for f(x) is considered more simple by this heuristic if it can
be solved).

ode_sol_simplicity() behaves identically to compare_ode_sol() when used as
a key function to min instead of a cmp function to sorted, but it only
needs to run once per solution, so it is much faster.  I also added a
trysolving key to disable attempting to solve the solution if you
already know it can't be done, (i.e., dsolve already tries unless it has
the simplify=False flag).  This makes finding the best solution given a
list of solutions that you already know are solved if they can be almost
instantaneous.

I replaced the use of compare_ode_sol in the code and all tests pass.

Also, replace map with a list comprehension in test_constantsimp.py.