refine: Let refine_Pow work harder to simplify powers of (-1)
Before, 'refine_Pow' would simplify (-1)**exp only if 'exp' as a whole could
be determined as even or odd.  A doctest like,

	>>> refine_Pow((-1)**(x+y), Assume(x, Q.even))
	(-1)**y

would fail.  This patch fixes that by treating powers of S.NegativeOne as a
special case.  Every term in the exponent is checked, and even terms are
simply discarded.  Odd terms are collected to an even quantity and then
discarded, possibly introducing a new term +1 in the exponent.  Finally,
a number N is replaced with N % 2.

We are now able to refine:

	>>> refine_Pow((-1)**(x+y+z), Assume(x, Q.odd) & Assume(z, Q.odd))
	(-1)**y
	>>> refine_Pow((-1)**(x+y+2), Assume(x, Q.odd))
	(-1)**(1 + y)
	>>> refine((-1)**(x+3))
	(-1)**(x+1)

These examples are implemented as doctests and regular tests.

Signed-off-by: Vinzent Steinberg <Vinzent.Steinberg@gmail.com>