1845: code quality test can see \r\n
When a file is opened in "r" mode all lines appear to
end in \n. You must open for reading (and writing) in
binary mode if you want to see (and correct) the line
endings. These changes have been made to the whitespace
fixer and the code quality tester. The same "b" mode
is used when opening for writing to keep files from being
written in a system-dependent format rather than with
the \n endings that are being enforced.