Implemented function for computing symmetric reduction
The new function is symmetrize(). It takes a multivariate
polynomial (not necessarily symmetric) and rewrites it in
terms of elementary symmetric polynomials. The output is
formed as a tuple with the symmetric decomposition and
the non-symmetric part.

Example:

In [1]: symmetrize(x**2 + y**2)
Out[1]:
⎛                2   ⎞
⎝-2⋅x⋅y + (x + y) , 0⎠

In [2]: symmetrize(x**2 + y**2, formal=True)
Out[2]:
⎛          2                         ⎞
⎝-2⋅s₂ + s₁ , 0, {s₁: x + y, s₂: x⋅y}⎠

In [3]: symmetrize(x**2 - y**2)
Out[3]:
⎛                2      2⎞
⎝-2⋅x⋅y + (x + y) , -2⋅y ⎠

In [4]: symmetrize(x**2 - y**2, formal=True)
Out[4]:
⎛          2      2                      ⎞
⎝-2⋅s₂ + s₁ , -2⋅y , {s₁: x + y, s₂: x⋅y}⎠

In [5]: Add(*_1).expand()
Out[5]:
 2    2
x  + y

In [6]: Add(*_3).expand()
Out[6]:
 2    2
x  - y