Implemented tools for counting real and complex roots
By default only real roots are counted, e.g.:

In [1]: count_roots(x**2 - 2)
Out[1]: 2

In [2]: count_roots(x**2 + 2)
Out[2]: 0

This is consistent with root isolation functions. One can also
specify an interval on the real line in which root counting is
performed, e.g.:

In [3]: count_roots(x**2 - 2, 0, oo)
Out[3]: 1

To make count_roots() switch to complex root counting algorithm
(Collins-Krandick), use a complex number for any of the lower
or the upper bound values. In the following example roots will
be counted in an interval [-2*I, 2*I] on the imaginary axis:

In [4]: count_roots(x**2 + 2, -2*I, 2*I)
Out[4]: 2

In [5]: count_roots(x**2 + 2, -2*I, 0)
Out[5]: 1

In [6]: count_roots(x**2 + 2, 0, 2*I)
Out[6]: 1

In general rectangular areas are supported, where the lower bound is
the south-western corner and the upper bound is the north-eastern
corner of the bounding rectangle. Roots are always counted including
the boundary.