1794: fast divisors (Tim Peters) and divisor_count
A less naive generator of divisors is included. I can take
credit only in making it sympy compatible and updating the
code a bit. Tim's amazing. ;-)

Since changes were made to make divisor enumeration faster,
a function divisor_count() was added which gives the number
of divisors e.g. 4 for n=6 (viz. 1, 2, 3, 6). One needs only
the factorization for this. Ref is included.

The factors in primefactors were being sorted twice so that
was changed and some tidying was done.

The limit is now included when factoring a number; that way if
someone wants to include everything up through a certain prime
they don't have to remember to add a 1 to that value.

The docstring for multiproduct was annotated a bit to make
it clear what is happening.

The built-in bisect replaces existing code.

Signed-off-by: Vinzent Steinberg <Vinzent.Steinberg@gmail.com>