1795: cartes(ian) product updated
cartes is now more useful in generating the Cartesian
product for an arbitrary number of sequences. Docstrings
show how sympy expressions can be generated, for example.

    >>> do=list(cartes([Mul, Add], [x, y], [2]))
    >>> for di in do:
    ... 	print di[0](*di[1:])
    ...
    2*x
    2*y
    2 + x
    2 + y

The variations (a.k.a permutations) has
been replaced with a version that uses iterators and breaks the
connection with cartes which was confusing and not documented well.

To round out the iterables, an option to compute subsets that allow
for repetition was added. The previous function itself was changed
to look more like the variations().

doc file updated.