1926: ode tweaks
1) I get nervous when I see direct string addressing...I think the
endswith/startswith methods might make it clearer what is being tested.

2) has_any_symbol() is an optimized routine to use when one is looking for
a symbol. It's the fastest method at present to answer the question about
whether a symbol appears in an expression.

3) <BLANKLINE> was deleted from a docstring

4) there were some minor text-edits made.

5) in odesimp,

a) everything is made into a list and at the very end it is returned as an
expression if there is only one element. I would prefer that it always
(like solve) return a list but I didn't try to change the existing behavior

b) a couple of assertions are added to validate the expectation and
existing code.

c) a loop over items in eq is made for renumbering. There are several cases
that were being skipped (but not generating an error) that involved lists.
To see them put the line "assert len(eq)==1" just before this loop.

6) as it's faster to test for inclusion in a set rather than a list, I made
some changes like that.

7) The sentence "The function is implemented recursively." got orphaned
from another docstring, I believe. So I deleted it.

8) the handling of the 1st_order case where the lhs was log(...) and the
rhs was 0 had Eq(newi.lhs.args[0]*C1, C1) but I think that
Eq(newi.lhs.args[0]/C1, C1) was intended, so I changed that and a test
that had begun to fail with the above changes then passed again.