1823: integral gets symbols property
>>> Integral(x, (y, 1, 2)).symbols
set([x])

There are times (like when solving equations) that you
need to know what symbols things depend on. .atoms(Symbol)
tells you too much. For the integral above, it would
return x and y. But the function really does not have
a surviving y in it after integration so should not be
considered a function of y.

All expr's have this property which defaults to atoms(Symbol);
any method that uses dummy variables (like Integral) should have
their own symbols method if one wants to not unclude the dummy
variables as variables on which the expr depends.