FCodePrinter: Always assume that loops could involve summation
This means that a loop for e.g. a matrix vector product y=Ax renders as:

do i=1,m
  do j=1,n
    y(i) = y(i) + A(i,j)*x(j)
  end do    ^
end do      |
            |....  Note that y has been added to the rhs expression.

In order for these loops to give correct results, the vector y must be
initialized correctly.  Here, we would need the fortran statement 'y=0.d0'
before the loop.  This must be taken care of by the CodeGen class.

Signed-off-by: Andy R. Terrel <andy.terrel@gmail.com>