FCodePrinter got method .set_assign_to()
FIXME: The assigment to a result variable should ideally be given not as a
printer parameter, but rather as a parmeter to .doprint().  However, this needs
changes throughout the Printer framework as .doprint() does not currently take
keyword arguments.   This should be fixed later.

For now the 'assign_to' setting is stored in the ``private'' Printer._settings
dictionary so that a setter method is apropriate.  We also need to check type
of the assigment variable as the loop functionality needs an Expr subclass to
work properly.

Signed-off-by: Andy R. Terrel <andy.terrel@gmail.com>