Codegen: setup of input/output arguments moved to Routine constructor
Before, the setup of arguments and return values had to be done prior to
calling the Routine constructor.  Now the routine constructor needs only
the name of the routine and the sympy expressions to calculate.  This should
help ensure consistency in how the Routine instances stores information.

The codegen() friendly function could be simplified considerably.

Updated tests that call the Routine constructor directly.

Signed-off-by: Andy R. Terrel <andy.terrel@gmail.com>