indexed: Change treatment of Idx(<literal integer constant>)
If an Idx instance is constructed based on a literal integer, the upper and
lower bounds should be preserved.

In relation to this, the *internal* structure of Idx instances with unspecified
range was also changed.  Before, an unspecified range would be stored as a tuple
(None, None), but this is redundant and also broke x.func(*x.args) == x.  Now,
Idx instances will only store the label as x.args[0], and nothing else.

Signed-off-by: Andy R. Terrel <andy.terrel@gmail.com>