test_external codegen: add tests that just inform about available compilers.
To test multpiple compilers and multiple languages, several command sequences
are defined.  Provided the compiler is present and behaves as expected, all
tests will be run on every available compiler.

In order to inform about which compilers are tested, this patch adds test
functions that simply call skip(<informative message>) for every compiler
that will not be tested.

Signed-off-by: Andy R. Terrel <andy.terrel@gmail.com>