Define __hash__ functions when __eq__ is overriden for Py3k support.
As hinted at in a bugreport[0] and implemented several years ago in the
py3k branch[1], in Python 3, overriding __eq__ without explicitly
defining __hash__ will cause __hash__ to be set to None.

I have merely defined __hash__ functions so that python2.7 rc2 will not
display any warnings (about this) when run with the -3 flag.  Since we
had not defined __hash__ methods previously, they are just calls to the
super-class's __hash__.

[0]: http://bugs.python.org/issue2235
[1]: http://svn.python.org/view?view=rev&revision=51454