Changed logic of Indexed, Idx and introduced IndexedElement
The Indexed module now has a better organisation in that the `stem' representing
the whole array is of class Indexed, while elements, e.g. A(i, j) are of class
IndexedElement.  This paves to way for storing shape information in the array
symbol as (optional?) arguments to Indexed constructor, and exploiting it in
expression manipulations.  Broadcast, for instance.

Signed-off-by: Andy R. Terrel <andy.terrel@gmail.com>