1979: atoms uses sets
Although sets are returned from atoms, the helper function uses
lists. If there are many symbols, all the duplicates get stored
and aren't removed until the list is converted to a set.

This commit changes the internal structure to sets; this makes
atoms faster for large expressions where the time matters; for
small expressions these changes appear to give about a 7% slow
down but for a small expression the overall time is small.