ccode: added doprint() capable of loops and top-level piecewise
The CCodeGen.doprint() takes assign_to as an optional keyword.  This
apporach is an alternative to having assign_to as a setting of the
printer.

Removed old _print_Piecewise(s, e) because branching should be done
outside loops for efficiency. we now also call fold_piecewise on
the expression.  Added a test for that.