Started a module tensor.index_methods for supportive functions to Indexed
Implemented the function get_indices(expr) which analyzes an expression
to determine the indices of it.  Repreated indices are interpreted as summation
indices, so they are not returned as indices of the expression.  Indices of
non commuting factors are kept separate from those of commuting factors.

Currently we support only Mul and Add expressions (including nested expressions).