Refactor as_real_imag() and _eval_expand_complex()
Now, the latter calls the former.  This has the advantage of keeping the
real and imaginary parts separate when they are calculated.  Previously,
as_real_imag() was using as_coefficient(I) to find the real and
imaginary parts.  But just because a term is not explicitly written as
a*I doesn't mean that is isn't imaginary.  Now the calculations that
were done in _eval_expand_complex(), which already explicitly knew the
real and imaginary parts, are done in as_real_imag(), keeping the
separate, and _eval_expand_complex() just calls as_real_imag() and
returns re_part + im_part*I.

This also fixes issue 1985.