1987: Integral(x, (x, None, y)) and Integral.diff
7 issues are addressed:

1) When there is an "evaluate at" integral like
the following that is differentiated, the lower
limit should not be included in the returned result:

Integral(x, (y, f(x))).diff(x) should be
= (x*y).subs(y, f(x)).diff(x)
= (x*f(x)).diff(x)
= x*f(x).diff(x) + f(x)

2) When there is a lower limit, the code was previously
making the wrong dummy variables substitution. It should
be

Integral(x, (y, L(x), U(x))).diff(x)
= (x*y).subs(y, U(x)).diff(x) - (x*y).subs(y,L(x)).diff(x)
= (x*U(x)).diff(x) - (x*L(x)).diff(x)
= x*dU/dx + U - (x*dL/dx + L)
= x*(dU/dx - dL/dx) + U - L

where x*y represent the evaluated integral. This is equivalent
to NOT computing the integral and doing the following instead:

= x.subs(y, U(x))*dU/dx + x.subs(y, L(x))*dL/dx +
  Integral(x.subs(y, t).diff(x).subs(t, y), (y, L(x), U(x))) <-+
= x*dU/dx + x*dL/dx +                                          |
  Integral(x.diff(x), (y, L(x), U(x)))                         |
= x*dU/dx + x*dL/dx +                                          |
  Integral(1, (y, L, U))                                       |
= x*dU/dx + x*dL/dx +                                          |
  U - L                                                        |
                                                               |
The masking of the dummy integration variable in this line is what
was previoiusly incorrect.

3) In order for this to work, subs has to sanitize the limits of
integrals that are stored in condensed format: Integral(x, x)
means Integral(t, (t, x)); the x in this form can (and must) be
modifiable for the differentiating under the integral sign. This
amounts to replacing limits like Tuple(x) with Tuple(x, x); a
comment about this is in the _eval_subs method. This only needs to
be done when a limit matches Tuple(old). Some tests were corrected
to change this behavior.

4) differentation is now multi-integral compliant.
When there is a double (or greater) number of Integrals like
Integral(A, (x, a, b), (y, c, d)) it is treated as

Integral(X, (x, a, b)) where X = Integral(A, (y, c, d)); the treatment
is the same as shown above but X replaces the integrand 'x'

= X.subs(y, U(x))*dU/dx + X.subs(y, L(x))*dL/dx +
  Integral(X.subs(y, t).diff(x).subs(t, y), (y, L(x), U(x)))
etc...

5) A new method for integrals, is_zero, was added to help this routine
and others. It does the quick test to see if an integral is identically
0 by seeing if there is a S.Zero integrand or two equal limits. No
simplification is attempted. It is just a quick test for zero on something
that doesn't autosimplify to zero.

6) All arguments of the integral should be sympified on instantiation
and a test was added for code uncovered by integration tests for the
integration of terms of an Add.

7) The transform method no longer converts single-arg limits to double
arg limits (Integral(x, x).transform(x, a*x) doesn't return a limit of
(x, x) now).

Some cleanup of the integrals.py file was also done.

use is_zero in eval_integral